Added field_names_for_merged_formsets options.
Each field in this option list should be the related_name of an fk
pointing to the controller's content_class.

It does not cause the related controller's formset to be merged, but
makes its form fields usable in the others fieldsets options.

For example, itbm.sites.UserController has
field_names_for_merged_formsets = ('profile',) because 'profile' is the
related_name of itbm.models.UserProfile.user field.

The two forms which are UserController and UserProfileController's first
formset "look" merged in the main user object form, by default.
UserProfile fields can now be used in UserController.fieldsets.

Next commit: secure and refactor